+++ oscarapi/serializers/basket.py
@@ -1,12 +1,12 @@
 import logging
 from rest_framework import serializers
-from rest_framework.fields import get_attribute
 
 from oscarapi.utils import (
     overridable,
     OscarModelSerializer,
     OscarHyperlinkedModelSerializer,
 )
+from oscarapi.serializers.fields import TaxIncludedDecimalField
 from django.utils.translation import ugettext as _
 from oscar.core.loading import get_model
 from decimal import Decimal
@@ -37,20 +37,6 @@ class OfferDiscountSerializer(serializers.Serializer):
 
 class VoucherDiscountSerializer(OfferDiscountSerializer):
     voucher = VoucherSerializer(required=False)
-
-
-class TaxIncludedDecimalField(serializers.DecimalField):
-    def __init__(self, excl_tax_field=None, excl_tax_value=None, *args, **kwargs):
-        self.excl_tax_field = excl_tax_field
-        self.excl_tax_value = excl_tax_value
-        super(TaxIncludedDecimalField, self).__init__(*args, **kwargs)
-
-    def get_attribute(self, instance):
-        if instance.is_tax_known:
-            return super(TaxIncludedDecimalField, self).get_attribute(instance)
-        if self.excl_tax_field:
-            return get_attribute(instance, (self.excl_tax_field, ))
-        return self.excl_tax_value
 
 
 class BasketSerializer(serializers.HyperlinkedModelSerializer):
